# [ENHANCEMENT] Optimize regexps with multiple prefixes