@pracucci 2024-04-22
@bboreham I would like to move forward with this PR. With the aim to help, I opened a PR against your branch to address my own comments: https://github.com/bboreham/prometheus/pull/3