@bboreham 2024-05-26
Update on this: the extra benchmark Marco added shows that small sets of alternates slow down with my change, so more work is needed.

Meanwhile, I discovered another approach, to sort the alternates during regexp compilation https://github.com/grafana/regexp/pull/11, which works well when the strings have some similarity.